# Searcher: make hot loop lock free