--- a/cmd/searcher/protocol/searcher.go
+++ b/cmd/searcher/protocol/searcher.go
@@ -216,6 +216,25 @@ func (fm FileMatch) MatchCount() int {
 	return count
 }
 
+func (fm *FileMatch) Limit(limit int) {
+	for i, cm := range fm.ChunkMatches {
+		l := len(cm.Ranges)
+		if l <= limit {
+			limit -= l
+			continue
+		}
+
+		cm.Ranges = cm.Ranges[:limit]
+		if limit > 0 {
+			fm.ChunkMatches = fm.ChunkMatches[:i+1]
+		} else {
+			fm.ChunkMatches = fm.ChunkMatches[:i]
+		}
+		fm.LimitHit = true
+		return
+	}
+}
+
 type ChunkMatch struct {
 	Content      string
 	ContentStart Location