@keegancsmith 2022-07-12
Using the searcher benchmarks is so much fun :) Hope you are enjoying it!