@camdencheek 2022-07-12
> Hope you are enjoying it!

Definitely am! This kind of work is a nice diversion 😄